Add an up-to-date test for msys-1.0.dll
This is not meant to provide a safeguard against hackers, but a safeguard
against a stale msys-1.0.dll which was not updated when installing into
an existing Git setup.

If /etc/msys-1.0.dll.md5 exists and does not match the md5 sum of
/bin/msys-1.0.dll (traditionally the most likely not-correctly-updated
file), it will complain loudly, read: in red letters.

If the md5 sum matches, the file will be deleted because the check is no
longer necessary (or even misleading if it fails in the future).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>